refactor: image type as shared across contexts.
