+++ vigil/embedding/business_logic/gateways/embedding_model.py
@@ -1,7 +1,7 @@
 from typing import Protocol
 
 from vigil.embedding.business_logic.models.embedded_track import Embedding
-from vigil.embedding.business_logic.models.image import Image
+from vigil.shared_kernel.models.image import Image
 
 
 class EmbeddingModel(Protocol):
